+++ lang/ca/validation.php
@@ -1,0 +1,18 @@
+<?php
+
+return [
+    'custom' => [
+        'address_postal_code' => [
+            'regex' => 'El codi postal ha de tenir només xifres (fins a 20).',
+        ],
+        'shipping_postal_code' => [
+            'regex' => 'El codi postal ha de tenir només xifres (fins a 20).',
+        ],
+        'installation_postal_code' => [
+            'regex' => 'El codi postal ha de tenir només xifres (fins a 20).',
+        ],
+        'postal_code' => [
+            'regex' => 'El codi postal ha de tenir només xifres (fins a 20).',
+        ],
+    ],
+];